// src/app/shared/enum/service-category-enum.ts
export enum ServiceCategoryEnum {
  Languages = 1,
  RoomType = 2,
  PropertyType = 3,
  TotalRooms = 4,
  TotalBeds = 5,
  TotalBathrooms = 6,
  TotalAdultGuest = 7,
  TotalChildGuest = 8,
  Amenities = 9,
  Highlights = 10,
  PlaceType = 11,
  BookingPreferences = 12,
  HouseRules = 13,
  CancellationPolicy = 14,
  ExperienceType = 15,
  BusinessKeywords = 16,
  VehicleType = 17,
  RoomOffers = 18,
  Neighbourhood = 19,
  PropertyFacilities = 20,
  FlightStoppage = 21,
  FlightFareType = 22,
  FlightAirlines = 23,
  FlightTicketClass = 24,
  TaxiType = 25,
  TaxiBookingType = 26,
  VendorType = 27,
  UserType = 28,
  UserStatus = 29,
  ServiceTypeStatus = 30,
  HostServiceStatus = 31,
  AttachmentStatus = 32,
  StayStatus = 33,
  TourItenaryStatus = 34,
  SightStatus = 35,
  FlightStatus = 36,
  TourDepartureStatus = 37,
  BookingStatus = 38,
  ReviewStatus = 39,
  AddStatus = 40,
  Fees = 41,
  Identity = 44,
  Sort = 53
}

export enum ServiceFeeSubCategory {
  CleaningFee = 157,
  RoomPrice = 150
}

export enum ServiceType {
  Experience = 2,
  UniqueHome = 3,
  Business = 6,
  Taxi = 7,
  Tour = 14
}